Minor fixes in Newcomer checks page (#12116)
* Minor fixes in Duplicate checker page

* Update app/webpacker/components/NewcomerChecks/DobChecker.jsx


* Added tooltip for merge button

---------
